Revert "Fixed Cyclic Dependency"
This reverts commit 3fc1e76d297f6a64c4ffdd2eee86fc20b38df788.
